refactor: GithubUserInfoResponse로 이름 변경 및 패키지 이동
