+++ backend/src/main/java/develup/infra/auth/github/dto/GithubAccessTokenRequest.java
@@ -1,4 +1,4 @@
-package develup.infra.auth.github;
+package develup.infra.auth.github.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
