+++ backend/src/main/java/develup/infra/auth/github/dto/GithubUserInfoResponse.java
@@ -1,4 +1,4 @@
-package develup.infra.auth.github;
+package develup.infra.auth.github.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -6,7 +6,7 @@ import develup.application.auth.OAuthUserInfo;
 import jakarta.annotation.Nullable;
 
 @JsonNaming(SnakeCaseStrategy.class)
-public record GithubUserInfo(
+public record GithubUserInfoResponse(
         Long id,
         String login,
         String avatarUrl,
